requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
cachetools==4.2.4
certifi==2022.6.15
charset-normalizer==2.1.0
decorator==4.4.2
discord-webhook==0.16.3
Django==4.0.5
django-constance==2.9.1
django-maintenance-mode==0.18.0
django-picklefield==3.1
django-simple-history==3.1.1
django-unicorn==0.49.2
gunicorn==20.1.0
idna==3.3
orjson==3.8.5
Pillow==9.2.0
pipdeptree==2.2.1
pydantic==1.9.1
PySocks==1.7.1
python-dotenv==0.20.0
python-fsutil==0.9.3
requests==2.28.1
shortuuid==1.0.11
soupsieve==2.3.2.post1
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.9
